tests/wpt/web-platform-tests/html/webappapis/scripting/processing-model-2/unhandled-promise-rejections/promise-rejection-events-onerror.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link rel="help" href="https://html.spec.whatwg.org/#runtime-script-errors">
<link rel="help" href="https://html.spec.whatwg.org/#unhandled-promise-rejections">
<script>
'use strict';
setup({
  allow_uncaught_exception: true
});
async_test(function(t) {
  var e = new Error('e');
  var e2 = new Error('e2');

  window.onerror = function (msg, url, line, col, error) {
    t.step(function() {
      assert_equals(msg, 'Uncaught Error: e2');
      assert_equals(error, e2);
    });
    t.done();
  };

  window.onrejectionhandled = function() {
    // This should cause onerror
    throw e2;
  };

  var p = Promise.reject(e);
  setTimeout(function() {
    setTimeout(t.step_func(function() {
      // This will cause onrejectionhandled
      p.catch(function() {});
    }), 0);
  }, 0);
}, 'Throwing inside an unhandledrejection handler invokes the error handler.');
</script>
